+++ lib/web_ui/lib/src/engine/skwasm/skwasm_impl/image.dart
@@ -64,7 +64,7 @@ class SkwasmImage extends SkwasmObjectWrapper<RawImage> implements ui.Image {
       final DomImageBitmap bitmap =
         (await (renderer as SkwasmRenderer).surface.renderPictures(
           <SkwasmPicture>[recorder.endRecording() as SkwasmPicture],
-        )).imageBitmaps.first;
+        )).imageBitmaps.single;
       final DomOffscreenCanvas offscreenCanvas =
         createDomOffscreenCanvas(bitmap.width.toDartInt, bitmap.height.toDartInt);
       final DomCanvasRenderingContextBitmapRenderer context =
@@ -75,8 +75,7 @@ class SkwasmImage extends SkwasmObjectWrapper<RawImage> implements ui.Image {
 
       // Zero out the contents of the canvas so that resources can be reclaimed
       // by the browser.
-      offscreenCanvas.width = 0;
-      offscreenCanvas.height = 0;
+      context.transferFromImageBitmap(null);
       return ByteData.view(arrayBuffer.toDart);
     } else {
       return (renderer as SkwasmRenderer).surface.rasterizeImage(this, format);
