+++ lib/web_ui/skwasm/library_skwasm_support.js
@@ -35,8 +35,8 @@ mergeInto(LibraryManager.library, {
               data.surface,
               data.callbackId, {
                 "imageBitmaps": data.imageBitmaps,
-                "rasterStart": data.rasterStart,
-                "rasterEnd": data.rasterEnd,
+                "rasterStartMilliseconds": data.rasterStart,
+                "rasterEndMilliseconds": data.rasterEnd,
               },
             );
             return;
